Fix Storage Policy creation bug
A REST Policy resource uses JSON, which can omit the bindings when empty.
This results in the Google API Client object having nil bindings,
instead of an empty array as we would get for a default value in Protobuf.
Fix the issue by making sure the gapi bindings are an array.
Add test coverage for Policy.